data schema 반영 코드 수정
