--- lib/data/book.dart
@@ -1,19 +1,0 @@
-class Book {
-  String bookName;
-  bool finish;
-  DateTime recentTime;
-  String imagePath;
-  double progress;
-  DateTime startTime;
-  String author;
-
-  Book({
-    required this.bookName,
-    required this.finish,
-    required this.recentTime,
-    required this.imagePath,
-    required this.progress,
-    required this.startTime,
-    required this.author,
-  });
-}
