+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:phopes/book_loader.dart';
+import 'package:phopes/isar_services.dart';
 import 'code_login_page.dart';
 import '/study_plan_page.dart';
 import 'student_home_page.dart';
@@ -11,23 +11,19 @@ import "update_page.dart";
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
-  runApp(const MyApp());
+  runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
   @override
   Widget build(BuildContext context) {
-    loadBooks();
     return MaterialApp(
       title: "login",
       theme: ThemeData(primarySwatch: Colors.blue),
-      initialRoute: '/study_plan',
+      initialRoute: '/first',
       routes: {
-        '/first': (context) => const FirstPage(),
-        '/main': (context) => const StudentHomePage(),
-        '/code': (context) => const CodeLoginPage(),
+        '/first': (context) => FirstPage(),
+        '/main': (context) => StudentHomePage(),
         '/study_plan': (context) => const StudyPlanPage(),
         // '/book_detail_page': (context) => BookDetailPage(),
         '/update': (context) => const UpdatePage()
