+++ pubspec.yaml
@@ -54,6 +54,7 @@ dependencies:
   firebase_auth: ^4.2.9
   isar: ^3.0.5
   cloud_firestore: ^4.4.4
+  isar_flutter_libs: ^3.0.5
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -77,22 +78,7 @@ flutter:
   uses-material-design: true
   assets:
     - repo/images/dog.png
-    - assets/images/books.jpg
-    - assets/images/b.jpg
-    - assets/q656.pdf
-    - assets/littleprince/book.json
-    - assets/littleprince/bookChapters.json
-    - assets/littleprince/bookChapter/1.json
-    - assets/littleprince/bookChapter/2.json
-    - assets/littleprince/bookChapter/3.json
-    - assets/littleprince/bookChapter/4.json
-    - assets/littleprince/bookChapter/5.json
-    - assets/littleprince/bookChapter/6.json
-    - assets/littleprince/bookChapter/7.json
-    - assets/littleprince/bookChapter/8.json
-    - assets/littleprince/bookChapter/9.json
-    - assets/littleprince/bookChapter/10.json
-    - assets/images/littlePrince.jpg
+    
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
